+++ example/middlewares.ts
@@ -1,7 +1,11 @@
 import createHttpError from "http-errors";
 import assert from "node:assert/strict";
 import { z } from "zod";
-import { Middleware, type Method } from "express-zod-api";
+import {
+  createCookieMiddleware,
+  Middleware,
+  type Method,
+} from "express-zod-api";
 
 export const authMiddleware = new Middleware({
   security: {
@@ -20,6 +24,20 @@ export const authMiddleware = new Middleware({
     assert.equal(token, "456", createHttpError(401, "Invalid token"));
     return { authorized: "Jane Doe" };
   },
+});
+
+/** @desc This middleware uses cookie as an input source and reads session from it */
+export const sessionMiddleware = new Middleware({
+  security: { type: "cookie", name: "session" },
+  input: z.object({ session: z.object({ token: z.string() }) }),
+  handler: async ({ input: { session } }) => ({ session }),
+});
+
+/** @desc This middleware provides setCookie() helper to context */
+export const cookieAssistingMiddleware = createCookieMiddleware({
+  httpOnly: true,
+  sameSite: "lax",
+  path: "/",
 });
 
 export const methodProviderMiddleware = new Middleware({