+++ express-zod-api/src/index.ts
@@ -8,6 +8,7 @@ export { getMessageFromError } from "./common-helpers";
 export { ensureHttpError } from "./result-helpers";
 export { BuiltinLogger } from "./builtin-logger";
 export { Middleware } from "./middleware";
+export { createCookieMiddleware } from "./cookie-middleware";
 export {
   ResultHandler,
   defaultResultHandler,