+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Version 28
 
+### v28.1.0
+
+- Added support for cookie handling:
+  - Cookie parsing can be enabled and configured in config (requires to install `cookie-parser`);
+  - `cookies` and `signedCookies` can be used as `inputSources` in config;
+  - `createCookieMiddleware()` creates a Middleware that exposes `setCookie()` and `clearCookie()` helpers into context
+    as well as the `getCookie()` one as an alternative to using cookies within `inputSources`;
+  - Documentation depicts request parameters when Middleware has `security` schema with `type: cookie`.
+
 ### v28.0.1
 
 - Adjusted the list of well-known headers, recognized by Documentation generator: